requirements.txt
gunicorn
flask
requests
python-dotenv==0.10.5